requirements.txt
fastapi==0.102.0
uvicorn[standard]==0.23.2
httpx==0.27.0
python-dotenv==1.0.0
tenacity==9.0.0
pydantic==2.9.0
aiofiles==23.2.1
ollama==0.6.1
pytest==8.3.5
pytest-asyncio==0.24.0
psycopg2-binary==2.9.10
sqlmodel==0.0.24
python-dotenv==1.1.1
sqlalchemy==2.0.44
pytest-mock==3.14.1
